Make backend test conversions explicit for strict MSVC
Preserve the existing float pupil coordinates and double GoogleTest comparisons while spelling their conversions explicitly. Avoid a shadowed rejected-value local. The strict GCC backend-test target and conversion/shadow diagnostics across 26 changed backend/render test translation units pass; no additional device execution was needed for these test-only spelling changes.

+++ tests/backend/kernel_camera_frame_test.cpp
@@ -261,8 +261,8 @@ void CheckCameraFrame(const std::string& artifact) {
             EXPECT_NEAR(output[36 + j], catalogue->second[j], 2e-6);
             // Project the observed physical screens back into their own frame;
             // a wrong yet orthonormal reference axis must fail this comparison.
-            EXPECT_NEAR(dot(24, 8 + 4 * j), output[33 + j], 4e-6);
-            EXPECT_NEAR(dot(28, 8 + 4 * j), output[36 + j], 4e-6);
+            EXPECT_NEAR(static_cast<double>(dot(24, 8 + 4 * j)), output[33 + j], 4e-6);
+            EXPECT_NEAR(static_cast<double>(dot(28, 8 + 4 * j)), output[36 + j], 4e-6);
         }
         const auto ray = relativity::PastDirectedCameraRay(*boosted, direction);
         const auto screen = relativity::ObserverScreenBasis(*boosted, direction);
